% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abv_data.R
\docType{data}
\name{abv_data}
\alias{abv_data}
\title{Alcohol content assumptions}
\format{A data table}
\source{
Natcen assumptions supplemented with market research data and figures from academic publications.
}
\usage{
abv_data
}
\description{
These are the expected percentages of alcohol that each beverage contains (alcohol by volume, ABV).
We use separate values for normal beer (4.4\%), strong beer (8.4\%), spirits (38\%), sherry (17\%), wine (12.5\%), and alcopops (4.5\%).
}
\keyword{datasets}
