% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alc_audit.R
\name{alc_audit}
\alias{alc_audit}
\title{Alcohol Use Disorder Identification Test (AUDIT)}
\usage{
alc_audit(data)
}
\arguments{
\item{data}{Data table - the health survey dataset}
}
\value{
\itemize{
\item audit1-10 - score for individual AUDIT questions
\item audit - full AUDIT score (1-40)
\item audit_c - AUDIT-C score (1-12)
\item audit_cat - categorisation of AUDIT; low risk (1-7), medium risk (8-15), higher risk (16-19), possible dependence (20-40)
\item audit_c_cat - categorisation of AUDIT-C; low risk (1-4), medium risk (5-7), higher risk (8-10), possible dependence (11-12)
}
}
\description{
Calculating the AUDIT and AUDIT-C scores.
}
\examples{

\dontrun{

data <- read_2016()
data <- clean_age(data)
data <- clean_demographic(data)
data <- alc_drink_now(data)
data <- alc_sevenday(data)
data <- alc_weekmean(data)

}

}
