% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alc_drink_freq.R
\name{alc_drink_freq}
\alias{alc_drink_freq}
\title{Drink frequency}
\usage{
alc_drink_freq(x)
}
\arguments{
\item{x}{Drinking frequency}
}
\value{
Returns a numeric translation (days/week) of the categorical drinking frequency variable.
\item{Almost every day}{7}
\item{Five or six days a week}{5.5}
\item{Three or four days a week}{3.5}
\item{Once or twice a week}{1.5}
\item{Once or twice a month}{0.375}
\item{Once every couple of months}{0.188}
\item{Once or twice a year}{0.029}
}
\description{
Convert the surveyed drink frequency to a measure of days per week on which drinking occurred.
}
\examples{

\dontrun{

alc_drink_freq(1)

}

}
