% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alc_drink_now_allages.R
\name{alc_drink_now_allages}
\alias{alc_drink_now_allages}
\title{Whether someone drinks and frequency of drinking}
\usage{
alc_drink_now_allages(data)
}
\arguments{
\item{data}{Data table - the Health Survey for England dataset.}
}
\value{
\itemize{
\item drinks_now (drinker, non_drinker).
\item drink_freq_7d (numerical variable calculated from categorical responses to how frequently someone drinks)
}
}
\description{
Clean the data on whether someone is a current drinker (drank in last 12 months)
and if they did drink how frequently they did so.
}
\details{
Combines information from adults and children. Someone is classed as a drinker even if they only reported
drinking once or twice a year. Any missing data is
supplemented by responses to if currently drinks or if always non-drinker.
}
\examples{

\dontrun{
data <- read_2016()
data <- clean_age(data)
data <- alc_drink_now_allages(data)
}

}
