% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alc_sevenday_adult.R
\name{alc_sevenday_adult}
\alias{alc_sevenday_adult}
\title{Alcohol consumption in last seven-days (adults)}
\usage{
alc_sevenday_adult(data, abv_data = hseclean::abv_data,
  alc_volume_data = hseclean::alc_volume_data)
}
\arguments{
\item{data}{Data table - the Health Survey for England dataset}

\item{abv_data}{Data table - our assumptions on the alcohol content of different beverages in (percent units / ml)}

\item{volume_data}{Data table - our assumptions on the volume of different drinks (ml).}
}
\value{
\itemize{
\item n_days_drink - number of days drank in last 7.
\item peakday - total units on heaviest drinking day
\item d7nbeer_units
\item d7sbeer_units
\item d7spirits_units
\item d7sherry_units
\item d7wine_units
\item d7pops_units
\item binge_cat (did_not_drink, binge, no_binge)
}
}
\description{
Processes the information from the questions on drinking in the last seven days - how many times drank and characteristics of the heaviest drinking day.
}
\details{
We estimate the number of
UK standard units of alcohol drunk on the heaviest drinking day by using the data on how many of what size measures of
different beverages were drunk, and combining this with our standard assumptions about beverage volume and alcohol content.

In 2007 new questions were added asking which glass size was used when wine was consumed.
Therefore the post HSE 2007 unit calculations are not directly comparable to previous years’ data.

MISSING DATA

Normally, if one of the constituent variables was missing, the whole variable would be marked as missing.
However, due to high missingness, we just assume any missing = 0, and so are likely to make underestimates.
}
\examples{

\dontrun{

data <- read_2001()
data <- clean_age(data)
data <- clean_demographic(data)
data <- alc_drink_now(data)
data <- alc_sevenday(data)

}

}
