% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alc_volume_data.R
\docType{data}
\name{alc_volume_data}
\alias{alc_volume_data}
\title{Beverage volume assumptions}
\format{A data table}
\source{
Natcen assumptions supplemented with market research data and figures from academic publications.
}
\usage{
alc_volume_data
}
\description{
These are the expected volumes (ml) of different beverage containers / serving sizes.
We use separate values for normal and strong beer (half pint 284ml, small can 330ml, large can 440ml, bottle 330ml),
spirits (serving 25ml),
sherry (serving 50ml), wine (small glass 125ml, standard glass 175ml, large glass 250ml, bottle 750ml),
and alcopops (small can 250ml, small bottle 275ml, large bottle 700ml).
}
\keyword{datasets}
