% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_economic_status.R
\name{clean_economic_status}
\alias{clean_economic_status}
\title{Economic status}
\usage{
clean_economic_status(data)
}
\arguments{
\item{data}{Data table - the Health Survey for England dataset.}
}
\value{
\itemize{
\item employ2cat: "employed", "unemployed"
\item nssec3_lab: "Managprof", "Intermediate", "Routinemanual", "Other". Social class/occupation based on the NS-SEC system. A four level variable for
managerial and professional occupations, intermediate occupations and routine and manual occupations. The fourth level is
'other' that includes anyone (young/old/unemployed) who do not have an occupation.
\item social_grade: "ABC1", "C2DE". 2 level social grade classification.
\item man_nonman: "Manual", "Non-manual", "Other"
\item activity_lstweek: "Managprof" (employed), "Intermediate" (employed), "Routinemanual" (employed), "Unemployed", "In education", "Looking after home or family", "Sickness, injury or disability", "Retired"
}
}
\description{
Process the data on employment and occupation.
}
\details{
The issues around using occupation-based social classifications for social survey research are discussed by
 \href{https://journals.sagepub.com/doi/full/10.1177/2059799116638003}{Connelly et al. 2016}. They advise using a range of alternative measures,
 but not creating new measures beyond what is already established.

The classifications considered are:

\itemize{
\item Employed / in paid work or not.
\item The \href{https://www.ons.gov.uk/methodology/classificationsandstandards/otherclassifications/thenationalstatisticssocioeconomicclassificationnssecrebasedonsoc2010}{NS-SEC measure}
which was constructed to measure the employment relations and conditions of occupations (i.e. it classifies people based on their employment occupation). It is
therefore not that good at classifying people who are not employed for various reasons.
\item The \href{https://humanrights.brightblue.org.uk/blog-1/2016/5/6/measuring-social-class}{NRS social grade system}. This measure is the one used in the
Tobacco and Alcohol Toolkit studies, but is not reported in the Health Survey for England. We create this variable by recategorising the NS-SEC 8 level variable.
This is important to facilitate the link of analysis to the Toolkit Study.
\item Manual vs. non-manual occupation. In the 2017 Tobacco control plan for England, there was a specific target to reduce the difference in rates of smoking
between people classified with a manual or non-manual occupation. We create this variable from the 3 level NS-SEC classification
by grouping Managerial and professional with intermediate occupations to give the non-manual group.
\item Economic status - retired / employed / unemployed.
\item Activity status for last week that adds more detail such as 'in education' and 'looking after home or family'.
}
}
\examples{

\dontrun{

data_2016 <- read_2016()

data_2016 <- clean_economic_status(data = data_2016)

}

}
