% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_income.R
\name{clean_income}
\alias{clean_income}
\title{Income}
\usage{
clean_income(data)
}
\arguments{
\item{data}{Data table - the Health Survey for England dataset.}
}
\value{
\itemize{
\item income5cat: quintiles of equivalised household income.
}
}
\description{
Process the data on income.
}
\details{
There are a few different options for classifying income - the need to have a measure that is
consistent across years of the Health Survey for England has led us to use equivalised income quintiles only.
Past SAPM modelling has used years of the Health Survey for England for which a continous variable for
equivalised income was provided. From this, we calculated our own income groups. In later years, this continuous
income variable is not available (except if we were to apply for a secure access version of the data).

In the past a measure of in poverty / not in poverty has been used, where the poverty threshold is defined as
60% of the median income for any year. For years in which we only have income quintiles available, it is not possible
to make an exact calculation of poverty. But it will coincide approximately with the lowest 2 income quintiles.

It would also be possible from the Health Survey for England to classify people as being in receipt of benefits or not.
This is not done currently, and would have to have some thought on how to
deal with the changing definitions of benefits over time.
}
\examples{

\dontrun{

data <- read_2004()
data <- clean_income(data = data)

}

}
