% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_years.R
\name{combine_years}
\alias{combine_years}
\title{Combine years of data}
\usage{
combine_years(data_list)
}
\arguments{
\item{data_list}{List of data tables to combine.}
}
\value{
Returns data table of combined data.
}
\description{
Combines years of data when provided as a list of data tables.
}
\examples{

\dontrun{

data_all <- combine_years(list(data1, data2, data3))

}

}
