% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_cat.R
\name{impute_cat}
\alias{impute_cat}
\title{Fills missing values of categorical variables with random draw from subgroup}
\usage{
impute_cat(
  data,
  var_names,
  strat_vars = c("year", "sex", "imd_quintile", "age_cat")
)
}
\arguments{
\item{data}{Data table - the Health Survey for England data}

\item{var_names}{Character vector - the variable names to be imputed (categorical variables only)}

\item{strat_vars}{Character vector - the variables by which to stratify the sample to create subgroups}
}
\value{
Returns an updated version of data in which the variables specified have had
their missing values imputed with random draws from the subgroups.
}
\description{
Replaces missing (NA) values with a random draw from the subgroup - 
the frequency of non-missing values within the subgroup therefore determines the result.
}
\details{
If not all NAs can be imputed with the fine scale starting amount of stratification,
imputation is attempted again, removing the stratification variable specified last.
}
\examples{

\dontrun{

library(hseclean)
library(data.table)
library(magrittr)

data <- read_2017(root = "/Volumes/Shared/")

data \%<>\%
  clean_age \%>\%
  clean_demographic \%>\% 
  clean_education \%>\%
  clean_economic_status \%>\%
  clean_family \%>\%
  clean_income \%>\%
  clean_health_and_bio \%>\%
  alc_drink_now_allages \%>\%
  alc_weekmean_adult \%>\%
  alc_sevenday_adult \%>\%
  alc_sevenday_child

data <- data[age >= 13, c("year", "age", "age_cat", "sex", "imd_quintile", "drinks_now", "drink_freq_7d", "total_units7_ch", "weekmean", "drinker_cat")]

data[is.na(drinks_now)]

data <- impute_cat(data, var_names = c("drinks_now"))

data[is.na(drinks_now)]

}

}
