% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_mean.R
\name{impute_mean}
\alias{impute_mean}
\title{Fills missing values with subgroup mean value \lifecycle{deprecated}}
\usage{
impute_mean(
  data,
  var_names,
  remove_zeros = FALSE,
  strat_vars = c("year", "sex", "imd_quintile", "age_cat")
)
}
\arguments{
\item{data}{Data table - the Health Survey for England data}

\item{var_names}{Character vector - the variable names to be imputed (numeric variables only)}

\item{remove_zeros}{Logical - should zeros be treated as missing data}

\item{strat_vars}{Character vector - the variables by which to stratify the subgroup means}
}
\value{
Returns an updated version of data in which the variables specified have had
their missing values imputed with the subgroup means.
}
\description{
Replaces any values < 0 with NA, calculates the subgroup mean,
then replaces missing values with the subgroup mean.
}
\details{
If not all NAs can be imputed with the fine scale starting amount of stratification,
imputation is attempted again, removing the stratification variable specified last.
}
\examples{

\dontrun{

data <- read_2001()
data <- clean_age(data)
data <- clean_demographic(data)
data <- impute_mean(data, var_names = c("d7many"))

}

}
