% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_capita_alc_for_upshift.R
\docType{data}
\name{per_capita_alc_for_upshift}
\alias{per_capita_alc_for_upshift}
\title{Per capita sales of alcohol for use in upshifting procedure from HMRC data
on duty receipts disaggregated by UK nation}
\format{
A data table
}
\source{
HMRC data on duty receipts disaggregated by UK nation: https://www.gov.uk/government/statistics/disaggregation-of-hmrc-tax-receipts
}
\usage{
per_capita_alc_for_upshift
}
\description{
Code to generate these data is in the "data-raw" folder of the hseclean package.
The year of duty take data used is 2018-19. Population counts are for 2018 are used.
For England and Wales, population counts were obtained from the
Office for National Statistics.
For Scotland, population counts were obtained from National Records Scotland.
}
\keyword{datasets}
