% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_capita_alc_for_upshift.R
\docType{data}
\name{per_capita_alc_for_upshift_scotland}
\alias{per_capita_alc_for_upshift_scotland}
\title{Per capita sales of alcohol for use in upshifting procedure from MESAS data on alcohol sales in Scotland.}
\format{
A data table
}
\source{
MESAS data on alcohol sales in Scotland: https://www.publichealthscotland.scot/publications/mesas-monitoring-report-2022/
}
\usage{
per_capita_alc_for_upshift_scotland
}
\description{
Code to generate these data is in the "data-raw" folder of the hseclean package.
The years data used is 2008 - 2019.
For Scotland, population counts were obtained from National Records Scotland.
}
\keyword{datasets}
