% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_counts.R
\docType{data}
\name{pop_counts}
\alias{pop_counts}
\title{Population sizes from 2001 to 2016}
\format{A data table (age, sex, year, imd_quintile, N).}
\source{
Observed population data provided under a data sharing agreement with the Office for National Statistics -
these data are provided in the same bundle as the deaths microdata.
}
\usage{
pop_counts
}
\description{
We use the observed data stratified by single years of age, sex and IMD quintile.
The data are held on the heta-study virtual machine and processed with the mort.tools R package.
}
\keyword{datasets}
