% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_summary.R
\name{prop_summary}
\alias{prop_summary}
\title{Summarise the percentages of people with certain levels of a variable}
\usage{
prop_summary(data, var_name, levels_1, levels_0, strat_vars = NULL)
}
\arguments{
\item{data}{Data table of individual characteristics.}

\item{var_name}{Character - the name of the variable to be summarised.}

\item{levels_1}{Character vector - the levels that should be assigned 1 in the binary variable}

\item{levels_0}{Character vector - the levels that should be assigned 0 in the binary variable}

\item{strat_vars}{Character vector - the variables by which to stratify the estimates}
}
\value{
Returns the estimate and standard error.
}
\description{
Uses the 'survey' package to estimate the percentages with uncertainty that accounts for survey sampling
error and survey design.
}
\examples{

\dontrun{

prop_smokers <- prop_summary(
  data = data,
  var_name = "cig_smoker_status",
  levels_1 = "current",
  levels_0 = c("former", "never"),
  strat_vars = c("sex", "imd_quintile")
)

}

}
