% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_2012.R
\name{read_2012}
\alias{read_2012}
\title{Read the Health Survey for England 2012}
\usage{
read_2012(
  root = c("X:/", "/Volumes/Shared/")[1],
  file =
    "HAR_PR/PR/Consumption_TA/HSE/Health Survey for England (HSE)/HSE 2012/UKDA-7480-tab/tab/hse2012ai.tab",
  select_cols = c("tobalc", "all")[1]
)
}
\arguments{
\item{root}{Character string - the root directory. This is the section of the file path to where the data is stored
that might vary depending on how the network drive is being accessed. The default is "X:/",
which corresponds to the University of Sheffield's X drive in the School of Health and Related Research.
Within the function, the root is pasted onto the front of the rest of the file path specified in the 'file' argument.
Thus, if root = NULL, then the complete file path is given in the 'file' argument.}

\item{file}{Character string - the file path and the name and extension of the file. The function has been
designed and tested to work with tab delimited files '.tab'. Files are read by the function [data.table::fread].}

\item{select_cols}{Character string - select either:
"all" - keep all variables in the survey data;
"tobalc" - keep a reduced set of variables associated with tobacco and alcohol consumption and a selected set of
survey design and socio-demographic variables that are needed for the functions within the hseclean package to work.}
}
\value{
Returns a data table.
}
\description{
Reads and does basic cleaning on the Health Survey for England 2012.
}
\section{Survey details}{

The HSE 2012 included a general population sample of adults and children, representative of
the whole population at both national and regional level. For the sample, 9,024 addresses
were randomly selected in 564 postcode sectors, issued over twelve months from January to
December 2013. Where an address was found to have multiple dwelling units, one dwelling
unit was selected at random and where there were multiple households at a dwelling unit, one
household was selected at random.

A total of 8,291 adults aged 16 and over and 2,043 children aged 0-15 were interviewed. A
household response rate of 64% was achieved for the core sample. Among the general
population sample, 5,470 adults and 1,203 children had a nurse visit.
}

\section{Weighting}{


For analyses at the individual level, the weighting variable to use is (wt_int). These weights are generated separately for adults and children:
\itemize{
\item for adults (aged 16 or more), the interview weights are a combination of the householdweight and a component which adjusts the sample to reduce bias from individual non-response within households;
\item for children (aged 0 to 15), the weights are generated from the household weights and the child selection weights – the selection weights correct for only including a maximum of two children in a household. The combined household and child selection weight were adjusted to ensure that the weighted age/sex distribution matched that of all children in co-operating households.
}
For analysis of children aged 0-15 in both the Core and the Boost sample, taking into account child selection only and not adjusting for non-response, the (wt_child) variable can be used. For analysis of children aged 2-15 in the only Boost sample the (wt_childb) variable can
}

\section{Missing values}{


\itemize{
\item -1 Not applicable: Used to signify that a particular variable did not apply to a given respondent
usually because of internal routing. For example, men in women only questions.
\item -8 Don't know, Can't say.
\item -9 No answer/ Refused
}
}

\section{How the data is read and processed}{

The data is read by the function [data.table::fread]. The 'root' and 'file' arguments are
pasted together to form the file path. The following are converted to NA:
c("NA", "", "-1", "-2", "-6", "-7", "-8", "-9", "-90", "-90.0", "-99", "N/A").
All variable names are converted to lower case.
The cluster and probabilistic sampling unit have the year appended to them.
Some renaming of variables is done for consistency with other years.
}

\examples{

\dontrun{

data_2012 <- read_2012("X:/",
"ScHARR/PR_Consumption_TA/HSE/HSE 2012/UKDA-7480-tab/tab/hse2012ai.tab")

}

}
