% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_2018.R
\name{read_2018}
\alias{read_2018}
\title{Read HSE 2018}
\usage{
read_2018(
  root = c("X:/", "/Volumes/Shared/"),
 
    file = "ScHARR/PR_Consumption_TA/HSE/Health Survey for England (HSE)/HSE 2018/tab/hse_2018_eul_22052020.tab"
)
}
\arguments{
\item{root}{Character - the root directory.}

\item{file}{Character - the file path and name.}
}
\value{
Returns a data table. Note that:
\itemize{
\item Missing data ("NA", "", "-1", "-2", "-6", "-7", "-8", "-9", "-90", "-90.0", "-99", "N/A") is replaced with NA.
\item All variable names are converted to lower case.
\item The cluster and probabilistic sampling unit have the year appended to them.
}
}
\description{
Reads and does basic cleaning on the Health Survey for England 2018.
}
\examples{

\dontrun{

data_2018 <- read_2018("X:/",
"ScHARR/PR_Consumption_TA/HSE/HSE 2018/UKDA-8334-tab/tab/hse2016_eul.tab")

}

}
