% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{select_data}
\alias{select_data}
\title{Select variables and apply filters}
\usage{
select_data(
  data,
  ages = 8:89,
  years = 2001:2017,
  keep_vars = c("age", "sex", "imd_quintile"),
  complete_vars = c("age", "sex", "imd_quintile")
)
}
\arguments{
\item{data}{Data table - the Health Survey for England dataset.}

\item{ages}{Integer vector - the ages in single years to retain (defaults to 8 to 89 years).}

\item{years}{Integer vector - the years in single years to retain (defaults to 2001 to 2016).}

\item{keep_vars}{Character vector - the names of the variables to keep.}

\item{complete_vars}{Character vector - the names of the variables on which the selection of complete cases will be based.}
}
\value{
Returns a reduced version of data
}
\description{
Selects the variables required for analysis and selects only the rows without missing data
for specified variables.
}
\examples{

\dontrun{

data <- select_data(data, keep_vars = c("age", "sex", "imd_quintile", "cig_smoker_status"))

}

}
