% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smk_amount.R
\name{smk_amount}
\alias{smk_amount}
\title{Characteristics of smoking \lifecycle{maturing}}
\usage{
smk_amount(data)
}
\arguments{
\item{data}{Data table - the Health Survey for England dataset.}
}
\value{
\itemize{
\item cigs_per_day - numeric (0+)
\item smoker_cat (non_smoker, 10_or_less, 11_to_20, 21_to_30, 31_or_more)
\item banded_consumption (non_smoker, light, moderate, heavy)
\item cig_type (non_smoker, hand rolled, machine rolled)
\item units_ROY_tob - numeric (0+) (years 2013+)
\item units_FM_cigs - numeric (0+) (years 2013+)
\item prop_handrolled - numeric (0-1) (years 2013+)
\item time_to_first_cig (non_smoker, less_than_5_minutes, five_to_thirty_minutes,
thirty_minutes_but_less_than_1_hour, one_hour_or_more)
}
}
\description{
Clean the variables that describe how much, what and to what level of addiction people smoke.
}
\details{
The main variable is the average number of cigarettes smoked per day. For adults, this is calculated
from questions about how many cigarettes are smoked typically on a weekday vs. a weekend 
(this is a weighted average to account for more weekdays in a week than weekends). For children,
this is based on asking how many cigarettes were smoked in the last week. Missing values are imputed as
the average amount smoked for an age, sex and IMD quintile subgroup.   

We categorise cigarette preferences based on the answer to 'what is the main type of cigarette smoked'. In
later years of the Health Survey for England, new questions are added from year 2013 that ask how many handrolled vs. machine rolled
cigarettes are smoked on a weekday vs. a weekend.    

We also categorise the amount smoked, and use information on the time from waking until smoking the first cigarette.
This latter variable has a high level of missingness. Together these categorical variables allow calculation of
 \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3307335/}{the heaviness of smoking index}.
}
\examples{

\dontrun{

library(hseclean)

data <- read_2017(root = "/Volumes/Shared/")
data <- clean_age(data)
data <- clean_demographic(data)
data <- smk_status(data)
data <- smk_amount(data)

}

}
