% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alc_drink_freq.R
\name{alc_drink_freq}
\alias{alc_drink_freq}
\title{Drink frequency}
\usage{
alc_drink_freq(x)
}
\arguments{
\item{x}{Drinking frequency}
}
\value{
Returns a numeric translation of the categorical drinking frequency variable.
}
\description{
Convert the surveyed drink frequency to a measure of days per week on which drinking occurred.
}
\examples{

\dontrun{

alc_drink_freq(1)

}

}
