% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_demographic.R
\name{clean_demographic}
\alias{clean_demographic}
\title{Demographic variables}
\usage{
clean_demographic(data)
}
\arguments{
\item{data}{Data table - the Health Survey for England dataset.}
}
\value{
\itemize{
\item ethnicity_4cat: 4 level variable (see above).
\item ethnicity_2cat: 2 level variable (white/nonwhite).
\item sex: m/f
\item imd_quintile
}
}
\description{
Processes demographic variables.
}
\details{
ETHNICITY
Previous SAPM modelling has used a simple white/non-white classification. The ONS recommend a harmonised ethnicity measure
for use in social surveys (\href{https://gss.civilservice.gov.uk/wp-content/uploads/2017/08/Ethnic-Group-June-17.pdf}{ONS, 2017}).
The use of ethnicity measures is also discussed in \href{https://journals.sagepub.com/doi/full/10.1177/2059799116642885}{Connelly et al. 2016},
who recommend testing the sensitivity of analyses to different specifications. We try to map the HSE categories to the ONS
recommended groups for England. However, over the years, the HSE is not clear or consistent in how they have categorised
chinese and arab as 'asian' or 'other'. In an attempt to harmonise, we have pooled the asian and other categories.
\itemize{
\item White (English, Irish, Scottish, Welsh, other European)
\item Mixed / multiple ethnic groups
\item Asian / Asian British (includes African-Indian, Indian, Pakistani, Bangladeshi),
plus Other ethnic group (includes Chinese, Japanese, Philippino, Vietnamese, Arab)
\item Black / African / Caribbean / Black British (includes Caribbean, African)
}
On the basis of this look at the data, the white/non-white classification does look appropriate, especially given
the likely limited sample sizes - so this 2 level variable has also been created.
}
\examples{

\dontrun{

data_2001 <- read_2001()

data_2001 <- clean_demographic(data = data_2001)

}

}
