% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_family.R
\name{clean_family}
\alias{clean_family}
\title{Family status}
\usage{
clean_family(data)
}
\arguments{
\item{data}{Data table - the Health Survey for England dataset.}
}
\value{
Returns an updated version of data with the new family variables:
\itemize{
\item 'kids' - number of children age <= 15 years in the household (0, 1, 2, 3+)
\item 'relationship_status' of the respondent (single, married, cohabit, sep_div_wid)
}
}
\description{
Process the data on the number of children in the household and the relationship
status of each respondent.
}
\details{
NUMBER OF CHILDREN AGE 0-15 YEARS IN THE HOUSEHOLD

Categorised into: 0, 1, 2, 3+. The problem with the Health Survey for England is that from 2015 onwards,
the number of children in the household is
not provided as this information could be identifiable (you can get it if you apply and pay for a secure dataset).
Therefore, for years 2015+, the number of children in the household is completely missing and needs to be imputed.

In the Scottish Health Survey, the number of children in the household is missing and needs to be imputed.


RELATIONSHIP STATUS

In previous versions of modelling (the SAPM alcohol binge model) relationship status has been described as
married/not-married. Here, we include more detail by using:
\itemize{
\item single
\item married, civil partnership or cohabiting
\item separated, divorced, widowed
}
}
\examples{

\dontrun{

library(hseclean)

#data <- read_2015(root = "/Volumes/Shared/")

storage_location <- "/Users/duncangillespie/Documents/HSE/"
data <- readRDS(paste0(storage_location, "HSE_2015.rds"))

data <- clean_age(data)
data <- clean_demographic(data)
data <- clean_education(data)
data <- clean_economic_status(data)
data <- clean_family(data)

}

}
