% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_health_and_bio.R
\name{clean_health_and_bio}
\alias{clean_health_and_bio}
\title{Health and biometric variables}
\usage{
clean_health_and_bio(data)
}
\arguments{
\item{data}{The Health Survey for England dataset.}
}
\value{
\itemize{
\item Returns a variable indicating the presence/absence of each
health condition (hse_cancer, hse_endocrine, hse_heart, hse_mental, hse_nervous, hse_eye, hse_ear, hse_respir, 
hse_disgest, hse_urinary, hse_skin, hse_muscskel, hse_infect, hse_blood).
\item height and weight.
\item BMI (numeric)
}
}
\description{
Cleans data on presence/absence of certain categories of health condition, and on height and weight.
}
\details{
HEATH CONDITIONS

There are a set of 14 categories of conditions that are ascertained consistently across all years of the Heath Survey for England.
 These are:
 \itemize{
 \item Cancer
 \item Endocrine or metabolic condition
 \item Mental health condition
 \item Nervous system condition
 \item Eye condition
 \item Ear condition
 \item Heart or circulatory system condition
 \item Respiratory condition
 \item Digestive condition
 \item Genito-urinary condition
 \item Skin condition
 \item Musculo-skeletal condition
 \item Infectious disease
 \item Blood and related organs condition
 }

 HEIGHT AND WEIGHT

 Height (cm) and weight (kg). Weight is estimated above 130kg. Missing values of height and weight are
 replaced by the mean height and weight for each age, sex and IMD quintile. 
 BMI is calculated according to kg / m^2.
}
\examples{

\dontrun{

data_2001 <- read_2001()

data_2001 <- clean_health_and_bio(data = data_2001)

}

}
