% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_surveyweights.R
\name{clean_surveyweights}
\alias{clean_surveyweights}
\title{Process the survey weights}
\usage{
clean_surveyweights(data, pop_data = hseclean::pop_counts)
}
\arguments{
\item{data}{Data table - the Health Survey for England dataset}

\item{pop_data}{Data table - population counts}
}
\value{
Returns data with survey weights updated
}
\description{
Process the survey weights
}
\examples{

\dontrun{

data <- clean_surveyweights(data)

}

}
