% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_surveyweights.R
\name{clean_surveyweights}
\alias{clean_surveyweights}
\title{Process the survey weights \lifecycle{maturing}}
\usage{
clean_surveyweights(data, pop_data)
}
\arguments{
\item{data}{Data table - the Health Survey for England dataset}

\item{pop_data}{Data table - population counts}
}
\value{
Returns data with survey weights updated
}
\description{
Makes new survey weights based on the difference between the HSE sample size and the actual population size 
by year, age, sex and IMD quintiles.
}
\examples{

\dontrun{

data <- clean_surveyweights(data, pop_data = stapmr::pop_counts)

}

}
