% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hseclean-package.R
\docType{package}
\name{hseclean-package}
\alias{hseclean}
\alias{hseclean-package}
\title{hseclean: Health Survey Data Wrangling}
\description{
Functions to read the data, rename, organise and process the variables required, impute missing data, and summarise data.
}
\author{
\strong{Maintainer}: Duncan Gillespie \email{duncan.gillespie@sheffield.ac.uk} (\href{https://orcid.org/0000-0003-3450-5747}{ORCID})

Authors:
\itemize{
  \item Laura Webster
  \item Colin Angus \email{c.r.angus@sheffield.ac.uk}
  \item Alan Brennan \email{a.brennan@sheffield.ac.uk}
}

}
\keyword{internal}
