% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_sim.R
\name{num_sim}
\alias{num_sim}
\title{Simulate single years from age/time category}
\usage{
num_sim(x)
}
\arguments{
\item{x}{Character - the age/time category e.g. "15-20".}
}
\value{
Returns an integer.
}
\description{
Used to overcome the loss of variables (age, time since quit) in single years in the Health Surveys for England 2015+.
}
\details{
Takes a category with beginning and end separated by "-" and randomly assigns each observation a number within the age category.
}
\examples{

num_sim("15-20")

}
