% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_2001.R
\name{read_2001}
\alias{read_2001}
\title{Read HSE 2001 \lifecycle{stable}}
\usage{
read_2001(
  root = c("X:/", "/Volumes/Shared/"),
  file = "ScHARR/PR_Consumption_TA/HSE/HSE 2001/UKDA-4628-tab/tab/hse01ai.tab"
)
}
\arguments{
\item{root}{Character - the root directory.}

\item{file}{Character - the file path and name.}
}
\value{
Returns a data table. Note that:
\itemize{
\item Missing data ("NA", "", "-1", "-2", "-6", "-7", "-8", "-9", "-90", "-90.0", "N/A") is replaced with NA.
\item All variable names are converted to lower case.
\item Each data point is assigned a weight of 1 as there is no weight variable supplied.
\item A single sampling cluster is assigned.
\item The probabilistic sampling unit have the year appended to them.
}
}
\description{
Reads and does basic cleaning on the Health Survey for England 2001.
}
\details{
A sample of the population living in private households. All persons living in the house, including those
under 2 years were eligible for inclusion. At addresses where there were more than two children under 16,
two children were selected at random. Information was obtained directly from persons aged 13 and
over. Information about children aged 0-12 was obtained from a parent, with the child present.

WEIGHTING

There is no weighted variable for household adult data.
For children under 16, the weighted variable Child_Wt should be used.

MISSING VALUES

\itemize{
\item -1 Not applicable: Used to signify that a particular variable did not apply to a given respondent
usually because of internal routing. For example, men in women only questions.
\item -2 Schedule not applicable: Used mainly for variables on the self-completions when the
respondent was not of the given age range, also used for children without legal guardians in the
home who could not participate in the nurse schedule.
\item -6 Schedule not obtained: Used to signify that a particular variable was not answered because the
respondent did not complete or agree to a particular schedule (i.e. nurse schedule or selfcompletions).
\item -7 Refused/ not obtained: Used only for variables on the nurse schedules, this code indicates that a
respondent refused a particular measurement or test or the measurement was attempted but not
obtained or not attempted.
\item -8 Don't know, Can't say.
\item -9 No answer/ Refused
}
}
\examples{

\dontrun{

data_2001 <- read_2001("X:/", "ScHARR/PR_Consumption_TA/HSE/HSE 2001/UKDA-4628-tab/tab/hse01ai.tab")

}

}
