% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_2017.R
\name{read_2017}
\alias{read_2017}
\title{Read HSE 2017}
\usage{
read_2017(root = c("X:/", "/Volumes/Shared/"),
  file = "ScHARR/PR_Consumption_TA/HSE/HSE 2017/UKDA-8488-tab/tab/hse17i_eul_v1.tab")
}
\arguments{
\item{root}{Character - the root directory.}

\item{file}{Character - the file path and name.}
}
\value{
Returns a data table. Note that:
\itemize{
\item Missing data ("NA", "", "-1", "-2", "-6", "-7", "-9", "-90", "-90.0", "N/A") is replace with NA,
except -8 ("don't know") as this is data.
\item All variable names are converted to lower case.
\item The cluster and probabilistic sampling unit have the year appended to them.
}
}
\description{
Reads and does basic cleaning on the Health Survey for England 2017.
}
\details{
The HSE 2017 sample comprised of a core general population sample. There was no boost sample in 2017.

The sample comprised 9,612 addresses selected at random in 534 postcode sectors, issued over twelve months from January to December 2017. Field work was completed in March 2018.

Adults and children were interviewed at households identified at the selected addresses. Up to four children in each household were selected to take part at random; up to two aged 2 to 12 and up to two aged 13 to 15.

A total of 7,997 adults aged 16 and over and 1,985 children aged 0-15 were interviewed, including 5,196 adults and 1,195 children who had a nurse visit.#'

From 2015 HSE data contains the 2015 English index of multiple deprivation, divided into quintiles.
}
\examples{

\dontrun{

data_2017 <- read_2017("X:/", "ScHARR/PR_Consumption_TA/HSE/HSE 2017/UKDA-8334-tab/tab/hse2016_eul.tab")

}

}
